+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/TxBody.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingVia #-}
@@ -65,7 +67,29 @@ module Cardano.Ledger.Dijkstra.TxBody (
   ),
   upgradeProposals,
   upgradeGovAction,
+  decodeGuards,
   DijkstraTxBodyRaw (..),
+  basicDijkstraTxBodyRaw,
+  inputsDijkstraTxBodyRawL,
+  outputsDijkstraTxBodyRawL,
+  feeDijkstraTxBodyRawL,
+  vldtDijkstraTxBodyRawL,
+  certsDijkstraTxBodyRawL,
+  withdrawalsDijkstraTxBodyRawL,
+  auxDataHashDijkstraTxBodyRawL,
+  mintDijkstraTxBodyRawL,
+  scriptIntegrityHashDijkstraTxBodyRawL,
+  collateralInputsDijkstraTxBodyRawL,
+  guardsDijkstraTxBodyRawL,
+  networkIdDijkstraTxBodyRawL,
+  collateralReturnDijkstraTxBodyRawL,
+  totalCollateralDijkstraTxBodyRawL,
+  referenceInputsDijkstraTxBodyRawL,
+  votingProceduresDijkstraTxBodyRawL,
+  proposalProceduresDijkstraTxBodyRawL,
+  currentTreasuryValueDijkstraTxBodyRawL,
+  treasuryDonationDijkstraTxBodyRawL,
+  subTransactionsDijkstraTxBodyRawL,
 ) where
 
 import Cardano.Ledger.Allegra.Scripts (invalidBeforeL, invalidHereAfterL)
@@ -76,33 +100,8 @@ import Cardano.Ledger.Babbage.TxBody (
   babbageSpendableInputsTxBodyF,
  )
 import Cardano.Ledger.BaseTypes (Network, StrictMaybe (..), fromSMaybe)
-import Cardano.Ledger.Binary (
-  Annotator,
-  DecCBOR (..),
-  Decoder,
-  EncCBOR (..),
-  Sized (..),
-  ToCBOR,
-  TokenType (..),
-  liftST,
-  mkSized,
-  peekTokenType,
- )
-import Cardano.Ledger.Binary.Coders (
-  Decode (..),
-  Density (..),
-  Encode (..),
-  Field,
-  Wrapped (..),
-  decode,
-  encode,
-  encodeKeyedStrictMaybe,
-  field,
-  fieldGuarded,
-  invalidField,
-  ofield,
-  (!>),
- )
+import Cardano.Ledger.Binary
+import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.Coin (Coin, decodePositiveCoin)
 import Cardano.Ledger.Conway (ConwayEra)
 import Cardano.Ledger.Conway.Core
@@ -200,6 +199,8 @@ instance
   , NFData (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   EqRaw (TxBody l DijkstraEra)
 
@@ -292,90 +293,104 @@ basicDijkstraTxBodyRaw SSubTx =
     mempty
     mempty
 
-instance (Typeable l, EraTxBody era) => DecCBOR (DijkstraTxBodyRaw l era) where
+instance
+  ( Typeable l
+  , EraTxBody era
+  , DecCBOR (Annotator (Tx SubTx era))
+  , OMap.HasOKey TxId (Tx SubTx era)
+  ) =>
+  DecCBOR (Annotator (DijkstraTxBodyRaw l era))
+  where
   decCBOR = withSTxBothLevels @l $ \sTxLevel ->
     decode $
       SparseKeyed
         "TxBodyRaw"
-        (basicDijkstraTxBodyRaw sTxLevel)
+        (pure $ basicDijkstraTxBodyRaw sTxLevel)
         (bodyFields sTxLevel)
-        requiredFields
+        (requiredFields sTxLevel)
     where
-      bodyFields :: STxBothLevels l era -> Word -> Field (DijkstraTxBodyRaw l era)
+      bodyFields :: STxBothLevels l era -> Word -> Field (Annotator (DijkstraTxBodyRaw l era))
       bodyFields sTxLevel = \case
-        0 -> field (inputsDijkstraTxBodyRawL .~) From
-        1 -> field (outputsDijkstraTxBodyRawL .~) From
-        2 | STopTx <- sTxLevel -> field (feeDijkstraTxBodyRawL .~) From
-        3 -> ofield (vldtDijkstraTxBodyRawL . invalidHereAfterL .~) From
+        0 -> fieldA (inputsDijkstraTxBodyRawL .~) From
+        1 -> fieldA (outputsDijkstraTxBodyRawL .~) From
+        2 | STopTx <- sTxLevel -> fieldA (feeDijkstraTxBodyRawL .~) From
+        3 -> ofieldA (vldtDijkstraTxBodyRawL . invalidHereAfterL .~) From
         4 ->
-          fieldGuarded
+          fieldAGuarded
             (emptyFailure "Certificates" "non-empty")
             OSet.null
             (certsDijkstraTxBodyRawL .~)
             From
         5 ->
-          fieldGuarded
+          fieldAGuarded
             (emptyFailure "Withdrawals" "non-empty")
             (null . unWithdrawals)
             (withdrawalsDijkstraTxBodyRawL .~)
             From
-        7 -> ofield (auxDataHashDijkstraTxBodyRawL .~) From
-        8 -> ofield (vldtDijkstraTxBodyRawL . invalidBeforeL .~) From
+        7 -> ofieldA (auxDataHashDijkstraTxBodyRawL .~) From
+        8 -> ofieldA (vldtDijkstraTxBodyRawL . invalidBeforeL .~) From
         9 ->
-          fieldGuarded
+          fieldAGuarded
             (emptyFailure "Mint" "non-empty")
             (== mempty)
             (mintDijkstraTxBodyRawL .~)
             From
-        11 -> ofield (scriptIntegrityHashDijkstraTxBodyRawL .~) From
+        11 -> ofieldA (scriptIntegrityHashDijkstraTxBodyRawL .~) From
         13
           | STopTx <- sTxLevel ->
-              fieldGuarded
+              fieldAGuarded
                 (emptyFailure "Collateral Inputs" "non-empty")
                 null
                 (collateralInputsDijkstraTxBodyRawL .~)
                 From
         14 ->
-          ofield
+          ofieldA
             (\x -> guardsDijkstraTxBodyRawL .~ fromSMaybe mempty x)
             (D decodeGuards)
-        15 -> ofield (networkIdDijkstraTxBodyRawL .~) From
+        15 -> ofieldA (networkIdDijkstraTxBodyRawL .~) From
         16
           | STopTx <- sTxLevel ->
-              ofield (collateralReturnDijkstraTxBodyRawL .~) From
+              ofieldA (collateralReturnDijkstraTxBodyRawL .~) From
         17
           | STopTx <- sTxLevel ->
-              ofield (totalCollateralDijkstraTxBodyRawL .~) From
+              ofieldA (totalCollateralDijkstraTxBodyRawL .~) From
         18 ->
-          fieldGuarded
+          fieldAGuarded
             (emptyFailure "Reference Inputs" "non-empty")
             null
             (referenceInputsDijkstraTxBodyRawL .~)
             From
         19 ->
-          fieldGuarded
+          fieldAGuarded
             (emptyFailure "VotingProcedures" "non-empty")
             (null . unVotingProcedures)
             (votingProceduresDijkstraTxBodyRawL .~)
             From
         20 ->
-          fieldGuarded
+          fieldAGuarded
             (emptyFailure "ProposalProcedures" "non-empty")
             OSet.null
             (proposalProceduresDijkstraTxBodyRawL .~)
             From
-        21 -> ofield (currentTreasuryValueDijkstraTxBodyRawL .~) From
+        21 -> ofieldA (currentTreasuryValueDijkstraTxBodyRawL .~) From
         22 ->
-          ofield
+          ofieldA
             (\x -> treasuryDonationDijkstraTxBodyRawL .~ fromSMaybe zero x)
             (D (decodePositiveCoin $ emptyFailure "Treasury Donation" "non-zero"))
+        23
+          | STopTx <- sTxLevel ->
+              fieldAA (subTransactionsDijkstraTxBodyRawL .~) (D decodeSubTransactions)
         n -> invalidField n
-      requiredFields :: [(Word, String)]
-      requiredFields =
-        [ (0, "inputs")
-        , (1, "outputs")
-        , (2, "fee")
-        ]
+      decodeSubTransactions :: Decoder s (Annotator (OMap TxId (Tx SubTx era)))
+      decodeSubTransactions =
+        decodeNonEmptySetLikeEnforceNoDuplicatesAnn
+          (flip (OMap.|>))
+          (\o -> (OMap.size o, o))
+      requiredFields :: STxBothLevels l era -> [(Word, String)]
+      requiredFields = \case
+        STopTx -> [(0, "inputs"), (1, "outputs"), (2, "fee")]
+        SSubTx -> [(0, "inputs"), (1, "outputs")]
+
       emptyFailure fieldName requirement =
         "TxBody: '" <> fieldName <> "' must be " <> requirement <> " when supplied"
 
@@ -446,6 +461,8 @@ deriving instance
   , NFData (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   NoThunks (TxBody l DijkstraEra)
 
@@ -454,6 +471,8 @@ deriving instance
   , NFData (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   Eq (TxBody l DijkstraEra)
 
@@ -462,6 +481,8 @@ deriving newtype instance
   , Eq (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   NFData (TxBody l DijkstraEra)
 
@@ -470,6 +491,8 @@ deriving instance
   , Eq (Tx SubTx DijkstraEra)
   , NFData (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   Show (TxBody l DijkstraEra)
 
@@ -478,6 +501,8 @@ pattern DijkstraTxBody ::
   , Eq (Tx SubTx DijkstraEra)
   , NFData (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   Set TxIn ->
   Set TxIn ->
@@ -596,6 +621,8 @@ pattern DijkstraSubTxBody ::
   , Eq (Tx SubTx DijkstraEra)
   , NFData (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   Set TxIn ->
   Set TxIn ->
@@ -696,9 +723,6 @@ type instance MemoHashIndex (DijkstraTxBodyRaw l DijkstraEra) = EraIndependentTx
 instance HashAnnotated (TxBody l DijkstraEra) EraIndependentTxBody where
   hashAnnotated = getMemoSafeHash
 
-instance (Typeable l, EraTxBody era) => DecCBOR (Annotator (DijkstraTxBodyRaw l era)) where
-  decCBOR = pure <$> decCBOR
-
 deriving via
   Mem (DijkstraTxBodyRaw l DijkstraEra)
   instance
@@ -707,6 +731,8 @@ deriving via
     , NFData (Tx SubTx DijkstraEra)
     , Show (Tx SubTx DijkstraEra)
     , EncCBOR (Tx SubTx DijkstraEra)
+    , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+    , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
     ) =>
     DecCBOR (Annotator (TxBody l DijkstraEra))
 
@@ -788,6 +814,8 @@ instance
   , NFData (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   EraTxBody DijkstraEra
   where
@@ -933,6 +961,8 @@ instance
   , NFData (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   AllegraEraTxBody DijkstraEra
   where
@@ -956,6 +986,8 @@ instance
   , NFData (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   MaryEraTxBody DijkstraEra
   where
@@ -996,6 +1028,8 @@ instance
   , NFData (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   AlonzoEraTxBody DijkstraEra
   where
@@ -1051,6 +1085,8 @@ instance
   , Eq (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   BabbageEraTxBody DijkstraEra
   where
@@ -1138,6 +1174,8 @@ instance
   , Eq (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   ConwayEraTxBody DijkstraEra
   where
@@ -1167,14 +1205,16 @@ guardsDijkstraTxBodyRawL =
         x@DijkstraSubTxBodyRaw {} -> \y -> x {dstbrGuards = y}
     )
 
-subTransactionsDijkstraTxBodyL :: Lens' (DijkstraTxBodyRaw TopTx era) (OMap TxId (Tx SubTx era))
-subTransactionsDijkstraTxBodyL = lens dtbrSubTransactions (\x y -> x {dtbrSubTransactions = y})
+subTransactionsDijkstraTxBodyRawL :: Lens' (DijkstraTxBodyRaw TopTx era) (OMap TxId (Tx SubTx era))
+subTransactionsDijkstraTxBodyRawL = lens dtbrSubTransactions (\x y -> x {dtbrSubTransactions = y})
 
 instance
   ( NFData (Tx SubTx DijkstraEra)
   , Eq (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   DijkstraEraTxBody DijkstraEra
   where
@@ -1182,7 +1222,7 @@ instance
   guardsTxBodyL = memoRawTypeL @DijkstraEra . guardsDijkstraTxBodyRawL
 
   {-# INLINE subTransactionsTxBodyL #-}
-  subTransactionsTxBodyL = memoRawTypeL @DijkstraEra . subTransactionsDijkstraTxBodyL
+  subTransactionsTxBodyL = memoRawTypeL @DijkstraEra . subTransactionsDijkstraTxBodyRawL
 
 -- | Decoder for decoding guards in a backwards-compatible manner. It peeks at
 -- the first element and if it's a credential, it decodes the rest of the